+++ plugins/partiql-memory/src/main/kotlin/org/partiql/plugins/memory/MemoryPlugin.kt
@@ -5,8 +5,9 @@ import org.partiql.spi.connector.Connector
 import org.partiql.spi.function.PartiQLFunction
 import org.partiql.spi.function.PartiQLFunctionExperimental
 
-class MemoryPlugin(val provider: MemoryCatalog.Provider) : Plugin {
-    override fun getConnectorFactories(): List<Connector.Factory> = listOf(MemoryConnector.Factory(provider))
+class MemoryPlugin(private val catalogs: Map<String, MemoryConnector>) : Plugin {
+
+    override fun getConnectorFactories(): List<Connector.Factory> = listOf(MemoryConnector.Factory(catalogs))
 
     @PartiQLFunctionExperimental
     override fun getFunctions(): List<PartiQLFunction> = emptyList()