+++ CHANGELOG.md
@@ -48,6 +48,7 @@ Thank you to all who have contributed!
   - Operators will be converted to function call. 
 - Changes the return type of `filter_distinct` to a list if input collection is list
 - Changes the `PartiQLValue` collections to implement Iterable rather than Sequence, allowing for multiple consumption.
+- **BREAKING** Moves PartiQLParserBuilder.standard().build() to be PartiQLParser.default().
 
 ### Deprecated